backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
numpy
scikit-learn
requests
